+++ src/ColorPicker/ColorPickerContainer.module.css
@@ -2,6 +2,8 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    margin-bottom: 50px;
+    border-bottom: 2px solid
 }
 
 .color-display {
