--- src/Food/Food.tsx
@@ -1,16 +1,0 @@
-
-function Food() {
-
-    const food1 = "Bulgogi Dish";
-    const food2 = "Korean Soup";
-
-    return (
-        <ul>
-            <h1>Favourite Food</h1>
-            <li>{food1}</li>
-            <li>{food2}</li>
-        </ul>
-    );
-}
-
-export default Food
