+++ src/Header/Header.module.css
@@ -6,6 +6,12 @@
   margin: 0;
 }
 
+.header h1 {
+  margin: 30px;
+  font-size: 3rem;
+  text-align: center;
+}
+
 .header ul {
   text-align: center;
 }
@@ -13,9 +19,4 @@
 .header ul li {
   display: inline;
   margin-right: 15px;
-}
-
-h1 {
-  margin: 50px;
-  font-size: 3rem;
 }
