+++ src/Home/Home.tsx
@@ -1,8 +1,12 @@
+import UserGreeting from "../UserGreeting/UserGreeting";
 
 function Home() {
     return (
         <div>
-            This is the home page, nothing to see here, go into other sub-pages XD
+            <UserGreeting isLoggedIn={true} username="Michael Scott" />
+            <h1>
+                This is the home page, nothing to see here, go into other sub-pages XD
+            </h1>
         </div>
     );
 }
