+++ src/List/List.module.css
@@ -1,5 +1,5 @@
 .list-category {
-    font-size: 2.5em;
+    font-size: 2em;
     font-weight: bold;
     color: hsl(0, 0%, 20%);
     margin-bottom: center;
@@ -10,11 +10,12 @@
 }
 
 .list-items {
-    font-size: 2.5em;
+    font-size: 1.5em;
     list-style: none;
     color: hsl(0, 0%, 25%);
     text-align: center;
     margin: 0;
+    margin-bottom: 60px;
 }
 
 .list-items li:hover {
