+++ src/Other/Other.module.css
@@ -1,0 +1,3 @@
+.header {
+    
+}
